+++ ldap2pg/config.py
@@ -314,6 +314,10 @@ class Configuration(dict):
         'postgres': {
             'dsn': '',
             'blacklist': ['pg_*', 'postgres'],
+            'databases_query': """
+            SELECT datname FROM pg_catalog.pg_database
+            WHERE datallowconn IS TRUE ORDER BY 1;
+            """.replace("\n" + ' ' * 12, "\n").strip(),
             # SQL Query to inspect roles in cluster. See
             # https://www.postgresql.org/docs/current/static/view-pg-roles.html
             # and
@@ -362,6 +366,7 @@ class Configuration(dict):
             secret=r'(?:password=|:[^/][^/].*@)',
         ),
         Mapping('postgres:blacklist', env=None),
+        Mapping('postgres:databases_query', env=None),
         Mapping('postgres:owners_query', env=None),
         Mapping('postgres:roles_query', env=None),
         Mapping('postgres:schemas_query', env=None),