+++ ldap2pg/script.py
@@ -51,6 +51,7 @@ def wrapped_main(config=None):
         acl_dict=config['acl_dict'],
         acl_aliases=config['acl_aliases'],
         blacklist=config['postgres']['blacklist'],
+        databases_query=config['postgres']['databases_query'],
         owners_query=config['postgres']['owners_query'],
         roles_query=config['postgres']['roles_query'],
         schemas_query=config['postgres']['schemas_query'],